requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
drf-yasg==1.21.7